Add unit test for basyx-python-sdk/basyx/aas/examples/data/_init_.py
